requirements/app.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements/app.txt requirements/app.in
#
arrow==0.12.0
attrs==17.4.0
certifi==2017.11.5        # via requests
chardet==3.0.4            # via requests
click==6.7
idna==2.6                 # via requests
oauthlib==2.0.6           # via requests-oauthlib
python-dateutil==2.6.1    # via arrow
requests-oauthlib==0.8.0
requests==2.18.4          # via requests-oauthlib
six==1.11.0               # via python-dateutil
urllib3==1.22             # via requests
